+++ src/filter.test.mjs
@@ -6,13 +6,13 @@ import { range } from 'itertools'
 
 test('filter base', async () => {
   const arr = [...range(6)]
-  const res = await filter(arr, async (x) => x % 2 === 0)
+  const res = await filter(arr, async (x) => x % 2 === 0, 2)
   expect(res).toEqual([0, 2, 4])
 })
 
 test('filter no async', async () => {
   const arr = [...range(6)]
-  const res = await filter(arr, (x) => x % 2 === 0)
+  const res = await filter(arr, (x) => x % 2 === 0, 2)
   expect(res).toEqual([0, 2, 4])
 })
 
@@ -27,7 +27,49 @@ test('filter concurrency', async () => {
     ds[x].resolve()
     await d.promise
     return x % 2 === 0
-  })
+  }, 2)
+  await ds[1].promise
+  expect(called[0]).toBe(1)
+  expect(called[1]).toBe(1)
+  expect(called[2]).toBe(0)
+  expect(called[3]).toBe(0)
+  expect(called[4]).toBe(0)
+  expect(called[5]).toBe(0)
+  d.resolve()
+  const res = await p
+  expect(res).toEqual([0, 2, 4])
+  expect(called[0]).toBe(1)
+  expect(called[1]).toBe(1)
+  expect(called[2]).toBe(1)
+  expect(called[3]).toBe(1)
+  expect(called[4]).toBe(1)
+  expect(called[5]).toBe(1)
+})
+
+test('filter infinite concurrency base', async () => {
+  const arr = [...range(6)]
+  const res = await filter(arr, async (x) => x % 2 === 0, Number.POSITIVE_INFINITY)
+  expect(res).toEqual([0, 2, 4])
+})
+
+test('filter infinite concurrency no async', async () => {
+  const arr = [...range(6)]
+  const res = await filter(arr, (x) => x % 2 === 0, Number.POSITIVE_INFINITY)
+  expect(res).toEqual([0, 2, 4])
+})
+
+test('filter infinite concurrency concurrency', async () => {
+  const arr = [...range(6)]
+  const called = {}
+  arr.forEach((v) => { called[v] = 0 })
+  const d = new Deferred()
+  const ds = arr.map(() => new Deferred())
+  const p = filter(arr, async (x) => {
+    called[x] += 1
+    ds[x].resolve()
+    await d.promise
+    return x % 2 === 0
+  }, Number.POSITIVE_INFINITY)
   await ds[5].promise
   expect(called[0]).toBe(1)
   expect(called[1]).toBe(1)
@@ -35,6 +77,48 @@ test('filter concurrency', async () => {
   expect(called[3]).toBe(1)
   expect(called[4]).toBe(1)
   expect(called[5]).toBe(1)
+  d.resolve()
+  const res = await p
+  expect(res).toEqual([0, 2, 4])
+  expect(called[0]).toBe(1)
+  expect(called[1]).toBe(1)
+  expect(called[2]).toBe(1)
+  expect(called[3]).toBe(1)
+  expect(called[4]).toBe(1)
+  expect(called[5]).toBe(1)
+})
+
+test('filter concurrency 1 base', async () => {
+  const arr = [...range(6)]
+  const res = await filter(arr, async (x) => x % 2 === 0)
+  expect(res).toEqual([0, 2, 4])
+})
+
+test('filter concurrency 1 no async', async () => {
+  const arr = [...range(6)]
+  const res = await filter(arr, (x) => x % 2 === 0)
+  expect(res).toEqual([0, 2, 4])
+})
+
+test('filter concurrency 1 concurrency', async () => {
+  const arr = [...range(6)]
+  const called = {}
+  arr.forEach((v) => { called[v] = 0 })
+  const d = new Deferred()
+  const ds = arr.map(() => new Deferred())
+  const p = filter(arr, async (x) => {
+    called[x] += 1
+    ds[x].resolve()
+    await d.promise
+    return x % 2 === 0
+  })
+  await ds[0].promise
+  expect(called[0]).toBe(1)
+  expect(called[1]).toBe(0)
+  expect(called[2]).toBe(0)
+  expect(called[3]).toBe(0)
+  expect(called[4]).toBe(0)
+  expect(called[5]).toBe(0)
   d.resolve()
   const res = await p
   expect(res).toEqual([0, 2, 4])
