+++ src/modern-async.mjs
@@ -10,8 +10,6 @@ export { default as Delayer } from './Delayer.mjs'
 export { default as every } from './every.mjs'
 export { default as filter } from './filter.mjs'
 export { default as filterGenerator } from './filterGenerator.mjs'
-export { default as filterLimit } from './filterLimit.mjs'
-export { default as filterSeries } from './filterSeries.mjs'
 export { default as find } from './find.mjs'
 export { default as findIndex } from './findIndex.mjs'
 export { default as findIndexLimit } from './findIndexLimit.mjs'
